ready to work on DAL

+++ Controller/FtpLoader.py
@@ -2,7 +2,8 @@ from Template.BaseClass import baseClass
 from utils.path import *
 import threading
 import logging
-from ftplib import FTP
+from ftplib import FTP, error_perm
+
 import socket
 import time
 import os
@@ -27,7 +28,7 @@ def setInterval(interval, times = -1):
 MONITOR_INTERVAL = 30
 class FtpLoader(baseClass):
 
-    __slots__ = ['host', 'port', 'ftp', 'ptr', 'max_attempts', 'waiting', 'logger','outDir']
+    __slots__ = ['host', 'port', 'ptr', 'max_attempts', 'waiting', 'logger','outDir']
 
     def __init__(self, host, port=21, outDir = None):
         self.logger = logging.getLogger(__name__)
@@ -84,6 +85,18 @@ class FtpLoader(baseClass):
             else:
                 self.ftp.close()
 
+    def ls(self, depth=0):
+        if depth > 10:
+            return ['depth > 10']
+        level = {}
+        for entry in (path for path in self.ftp.nlst() if path not in ('.', '..')):
+            try:
+                self.ftp.cwd(entry)
+                level[entry] = self.ls(self.ftp, depth + 1)
+                self.ftp.cwd('..')
+            except error_perm:
+                level[entry] = None
+        return level
     def getFileSize(self, file):
         return self.ftp.size(file)
 
@@ -102,6 +115,12 @@ class FtpLoader(baseClass):
             self.ftp.retrbinary('RETR %s' % filename, filePtr.write, rest=filePtr.tell())
         os.chdir(self.cwd)
         return res
+    def checkIfFileExist(self, fName):
+        files_list = self.ls()
+        for file, subfiles in files_list.items() :
+            if fName == file :
+                return True
+        return False
 
     def downloadFileItem(self, dst_filename):
 
@@ -115,34 +134,40 @@ class FtpLoader(baseClass):
         except:
             return False
         if conn:
-            with open(local_filename, 'w+b') as filePtr:
-                    self.ptr = filePtr.tell()
-                    dst_filesize = self.getFileSize(dst_filename)
-                    mon = self.monitor(filePtr)
-                    downloaded = ''
-                    while dst_filesize > filePtr.tell():
-                        try:
-                            downloaded = self.openDlStream(self.outDir, filePtr, dst_filename)
-                            if downloaded is None :
-                                break
-                        except:
-                            self.max_attempts -= 1
-                            if self.max_attempts == 0:
-                                mon.set()
-                                self.logger.exception('')
-                                raise
-                            self.waiting = True
-                            self.logger.info('waiting 30 sec...')
-                            time.sleep(30)
-                            self.logger.info('reconnect')
 
-                    if not downloaded.startswith('226 Transfer complete'):
-                        self.logger.error('Downloaded file {0} is not full.'.format(dst_filename))
-                        # os.remove(local_filename)
-                        return None
-                    mon.set() #stop monitor
-                    self.ftp.close()
-                    return True
+            if self.checkIfFileExist(dst_filename) :
+                with open(local_filename, 'w+b') as filePtr:
+                        self.ptr = filePtr.tell()
+                        dst_filesize = self.getFileSize(dst_filename)
+                        mon = self.monitor(filePtr)
+                        downloaded = ''
+                        while dst_filesize > filePtr.tell():
+                            try:
+                                downloaded = self.openDlStream(self.outDir, filePtr, dst_filename)
+                                if downloaded is None :
+                                    break
+                            except:
+                                self.max_attempts -= 1
+                                if self.max_attempts == 0:
+                                    mon.set()
+                                    self.logger.exception('')
+                                    raise
+                                self.waiting = True
+                                self.logger.info('waiting 30 sec...')
+                                time.sleep(30)
+                                self.logger.info('reconnect')
+
+                        if not downloaded.startswith('226 Transfer complete'):
+                            self.logger.error('Downloaded file {0} is not full.'.format(dst_filename))
+                            # os.remove(local_filename)
+                            return None
+                        mon.set() #stop monitor
+                        self.ftp.close()
+                        return True
+
+            else :
+                self.logger.error("the file {} does not exist in the FTP server. Check mispelling".format(dst_filename))
+                return None
